feat(fuzz): userId.fromObject
